Move cloud request building code to zeta2 + other misc changes
