+++ crates/zeta2/src/zeta2.rs
@@ -1,9 +1,14 @@
-use std::{ops::Range, sync::Arc};
-
-use gpui::{App, Entity, EntityId, Task, prelude::*};
-
+use cloud_llm_client::predict_edits_v3::{self, Signature};
 use edit_prediction::{DataCollectionState, Direction, EditPrediction, EditPredictionProvider};
+use edit_prediction_context::{
+    DeclarationId, EditPredictionContext, EditPredictionExcerptOptions, SyntaxIndex,
+    SyntaxIndexState,
+};
+use gpui::{App, Entity, EntityId, Task, prelude::*};
 use language::{Anchor, ToPoint};
+use language::{BufferSnapshot, Point};
+use std::collections::HashMap;
+use std::{ops::Range, sync::Arc};
 
 pub struct Zeta2EditPredictionProvider {
     current: Option<CurrentEditPrediction>,
@@ -151,4 +156,117 @@ impl EditPredictionProvider for Zeta2EditPredictionProvider {
 
         Some(current_prediction.prediction)
     }
+}
+
+pub fn make_cloud_request_in_background(
+    cursor_point: Point,
+    buffer: BufferSnapshot,
+    events: Vec<predict_edits_v3::Event>,
+    can_collect_data: bool,
+    diagnostic_groups: Vec<predict_edits_v3::DiagnosticGroup>,
+    git_info: Option<cloud_llm_client::PredictEditsGitInfo>,
+    excerpt_options: EditPredictionExcerptOptions,
+    syntax_index: Entity<SyntaxIndex>,
+    cx: &mut App,
+) -> Task<Option<predict_edits_v3::PredictEditsRequest>> {
+    let index_state = syntax_index.read_with(cx, |index, _cx| index.state().clone());
+    cx.background_spawn(async move {
+        let index_state = index_state.lock().await;
+        EditPredictionContext::gather_context(cursor_point, &buffer, &excerpt_options, &index_state)
+            .map(|context| {
+                make_cloud_request(
+                    context,
+                    events,
+                    can_collect_data,
+                    diagnostic_groups,
+                    git_info,
+                    &index_state,
+                )
+            })
+    })
+}
+
+pub fn make_cloud_request(
+    context: EditPredictionContext,
+    events: Vec<predict_edits_v3::Event>,
+    can_collect_data: bool,
+    diagnostic_groups: Vec<predict_edits_v3::DiagnosticGroup>,
+    git_info: Option<cloud_llm_client::PredictEditsGitInfo>,
+    index_state: &SyntaxIndexState,
+) -> predict_edits_v3::PredictEditsRequest {
+    let mut signatures = Vec::new();
+    let mut declaration_to_signature_index = HashMap::default();
+    let mut referenced_declarations = Vec::new();
+    for snippet in context.snippets {
+        let parent_index = snippet.declaration.parent().and_then(|parent| {
+            add_signature(
+                parent,
+                &mut declaration_to_signature_index,
+                &mut signatures,
+                index_state,
+            )
+        });
+        let (text, text_is_truncated) = snippet.declaration.item_text();
+        referenced_declarations.push(predict_edits_v3::ReferencedDeclaration {
+            text: text.into(),
+            text_is_truncated,
+            signature_range: snippet.declaration.signature_range_in_item_text(),
+            parent_index,
+            score_components: snippet.score_components,
+            signature_score: snippet.scores.signature,
+            declaration_score: snippet.scores.declaration,
+        });
+    }
+
+    let excerpt_parent = context
+        .excerpt
+        .parent_declarations
+        .last()
+        .and_then(|(parent, _)| {
+            add_signature(
+                *parent,
+                &mut declaration_to_signature_index,
+                &mut signatures,
+                index_state,
+            )
+        });
+
+    predict_edits_v3::PredictEditsRequest {
+        excerpt: context.excerpt_text.body,
+        referenced_declarations,
+        signatures,
+        excerpt_parent,
+        // todo!
+        events,
+        can_collect_data,
+        diagnostic_groups,
+        git_info,
+    }
+}
+
+fn add_signature(
+    declaration_id: DeclarationId,
+    declaration_to_signature_index: &mut HashMap<DeclarationId, usize>,
+    signatures: &mut Vec<Signature>,
+    index: &SyntaxIndexState,
+) -> Option<usize> {
+    if let Some(signature_index) = declaration_to_signature_index.get(&declaration_id) {
+        return Some(*signature_index);
+    }
+    let Some(parent_declaration) = index.declaration(declaration_id) else {
+        log::error!("bug: missing parent declaration");
+        return None;
+    };
+    let parent_index = parent_declaration.parent().and_then(|parent| {
+        add_signature(parent, declaration_to_signature_index, signatures, index)
+    });
+    let (text, text_is_truncated) = parent_declaration.signature_text();
+    let signature_index = signatures.len();
+    signatures.push(Signature {
+        text: text.into(),
+        text_is_truncated,
+        parent_index,
+    });
+    declaration_to_signature_index.insert(declaration_id, signature_index);
+    Some(signature_index)
 }
